Added some comments and deleted useless files
