+++ src/math_markup/mod.rs
@@ -1,12 +1,10 @@
 //! Utilities to deal with math markup.
 
 mod preferred_markup;
-// mod typst_base;
+mod typst_base;
 mod typst_main;
-mod typst_rebase;
-mod typst_stuff;
 
 pub(crate) use preferred_markup::{get_preferred_markup, set_default_math_markup};
 // pub(crate) use typst_base::TYPST_BASE;
+pub(crate) use typst_base::typst_render;
 pub(crate) use typst_main::{catch_typst_message, typst};
-pub(crate) use typst_rebase::typst_render;
