+++ src/math_markup/typst_base.rs
@@ -1,70 +1,23 @@
 use comemo::Prehashed;
+use image::ImageDecoder;
 use once_cell::sync::Lazy;
-use std::convert::TryInto;
-use std::sync::Arc;
-use typst::diag::{FileResult, SourceError};
-use typst::eval::Library;
-use typst::file::FileId;
-use typst::font::{Font, FontBook};
-use typst::geom::Size;
-use typst::syntax::Source;
-use typst::util::Bytes;
+use serenity::futures::lock::Mutex;
+use std::cell::RefCell;
+use std::io::Cursor;
+use std::sync::OnceLock;
+use typst::diag::{FileError, FileResult, Severity, SourceDiagnostic, SourceResult};
+use typst::eval::Tracer;
+use typst::foundations::{Bytes, Datetime};
+use typst::layout::{Abs, Axes};
+use typst::syntax::{FileId, PackageSpec, PackageVersion, Source};
+use typst::text::{Font, FontBook};
+use typst::Library;
 
-pub(crate) trait Preamble {
-    fn preamble(&self) -> String;
-}
-
-//TODO, allow changing the customisation
-#[derive(Debug, Clone, Copy)]
-pub(crate) enum PageSize {
-    Auto,
-}
-
-impl Preamble for PageSize {
-    fn preamble(&self) -> String {
-        match &self {
-            PageSize::Auto => "#set page(width: auto, height: auto, margin: 10pt)\n".to_string(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub(crate) enum Theme {
-    Dark,
-}
-
-impl Preamble for Theme {
-    fn preamble(&self) -> String {
-        match self {
-            //             Theme::Light => "
-            // #let bg = rgb(219, 222, 225)
-            // #let fg = rgb(49, 51, 56)
-            // "
-            //            .to_string(),
-            Theme::Dark => "
-#let fg = rgb(219, 222, 225)
-#let bg = rgb(49, 51, 56)"
-                .to_string(),
-        }
-    }
-}
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct CustomisePage {
-    pub(crate) page_size: PageSize,
-    pub(crate) theme: Theme,
-}
-
-impl Preamble for CustomisePage {
-    fn preamble(&self) -> String {
-        self.page_size.preamble() + self.theme.preamble().as_str()
-    }
-}
-
-pub(crate) struct TypstEssentials {
-    library: Prehashed<Library>,
-    fontbook: Prehashed<FontBook>,
-    fonts: Vec<Font>,
-    choices: CustomisePage,
+//Needed for reading files for packages
+#[derive(Clone)]
+struct FileEntry {
+    bytes: Bytes,
+    source: Option<Source>,
 }
 
 fn get_fonts() -> Vec<Font> {
@@ -78,146 +31,267 @@ fn get_fonts() -> Vec<Font> {
         })
         .collect()
 }
+/*
+To explain the package layout; because this might not be obvious
+packages/
+    {package_name}-{version}/
+        entry_point.typ -> The important file
+        LICENSE -> I need to check if this is legal, milord
+        typst.toml -> tells us where the entrypoint is
 
-impl TypstEssentials {
+I do not plan on doing stuff that involves downloading packages as needed, but if needed, the packages can be downloaded from
+https://packages.typst.org/{namespace}/{name}-{version}.tar.gz
+namespace will be preview, as per typst 0.10.0; the tar file should just be extracted into the folder
+*/
+
+impl FileEntry {
+    fn source(&mut self, id: FileId) -> FileResult<Source> {
+        // Fallible `get_or_insert`.
+        let src = match &self.source {
+            Some(e) => e,
+            None => {
+                let contents =
+                    std::str::from_utf8(&self.bytes).map_err(|_| FileError::InvalidUtf8)?;
+                // Defuse the BOM!
+                let contents = contents.trim_start_matches('\u{feff}');
+                let source = Source::new(id, contents.into());
+                self.source.insert(source)
+            }
+        };
+        Ok(src.clone())
+    }
+}
+
+pub(crate) struct TypstRendered {
+    library: Prehashed<Library>,
+    fontbook: Prehashed<FontBook>,
+    fonts: Vec<Font>,
+    files: RefCell<Vec<(FileId, FileEntry)>>, //This used to be a hashmap, but bruh, there are like 3 to 5 files here maximum, its easier to traverse a list; RefCell is needed because I need to mutate this sometimes.
+    source: Option<Source>,
+}
+impl TypstRendered {
     pub(crate) fn new() -> Self {
         let fonts = get_fonts();
-
         Self {
-            library: Prehashed::new(typst_library::build()),
+            library: Prehashed::new(Library::build()),
             fontbook: Prehashed::new(FontBook::from_fonts(&fonts)),
             fonts,
-            choices: CustomisePage {
-                page_size: PageSize::Auto,
-                theme: Theme::Dark,
-            },
+            files: RefCell::new(Vec::new()),
+            source: None,
         }
     }
-}
+    fn from_file_id(&self, file_id: &FileId) -> Option<FileEntry> {
+        self.files
+            .take()
+            .iter()
+            .find_map(|(a, b)| if a == file_id { Some(b.clone()) } else { None })
+    }
+    fn file_entry(&self, id: &FileId) -> FileResult<FileEntry> {
+        //Check if file is already in vector
+        if let Some(a) = self.from_file_id(&id) {
+            return Ok(a);
+        }
 
-impl Preamble for TypstEssentials {
-    fn preamble(&self) -> String {
-        self.choices.preamble()
-            + "
-#set text(
-  font: (
-    \"EB Garamond 12\",
-    \"DejaVu Sans Mono\"
-  ),
-  size: 18pt,
-  number-type: \"lining\",
-  number-width: \"tabular\",
-  weight: \"regular\",
-);
-// set a size that more closely approximates the zoom-in on Discord
+        //Check if it is a package, should be infallible, but yeah
+        let p = id
+            .package()
+            .ok_or(FileError::NotFound(id.vpath().as_rootless_path().into()))?;
+
+        let mut dir = std::path::PathBuf::new();
+        dir.push("packages");
+        let package_folder = format!(
+            "{}-{}.{}.{}",
+            p.name, p.version.major, p.version.minor, p.version.patch
+        );
+        dir.push(&package_folder);
+
+        let path = id.vpath().resolve(dir.as_path()).unwrap();
+        let contents = std::fs::read(&path).map_err(|error| FileError::from_io(error, &path))?;
+
+        let fe = FileEntry {
+            bytes: contents.into(),
+            source: None,
+        };
+        self.files.borrow_mut().push((id.clone(), fe));
+        // self.files.push((id.clone(), fe));
+        //Should return the above file_entry, and there should be an easy way to do it, but i dont know;
+
+        Ok(self.from_file_id(&id).unwrap().clone())
+    }
+    fn preamble() -> String {
+        let imports = r#"
+            #import "@preview/whalogen:0.1.0": *
+        "#;
+        let theme = r#"
+            #let fg = rgb(219, 222, 225)
+            #let bg = rgb(49, 51, 56)
+        "#;
+        let page_size = "#set page(width: auto, height: auto, margin: 10pt)\n";
+        let text_options = r#"
+            #set text(
+                font: (
+                    "EB Garamond 12",
+                    "DejaVu Sans Mono"
+                      ),
+                size: 18pt,
+                number-type: "lining",
+                number-width: "tabular",
+                weight: "regular",
+            );
+            // set a size that more closely approximates the zoom-in on Discord
 
 
-// Preamble starts here
+            // Preamble starts here
+            #set page(
+              fill: bg
+            )
+            #set text(
+              fill: fg,
+            )
+        "#;
+        let math_remaining = r#"
+            #let infty = [#sym.infinity];
 
-#set page(
-  fill: bg
-)
-#set text(
-  fill: fg,
-)
+            #let di = [#math.dif i];
+            #let du = [#math.dif u];
+            #let dr = [#math.dif r];
+            #let ds = [#math.dif s];
+            #let dt = [#math.dif t];
+            #let dx = [#math.dif x];
+            #let dx = [#math.dif x];
+            #let dy = [#math.dif y];
+            #let dz = [#math.dif z];
 
-#let infty = [#sym.infinity];
+            #let int = [#sym.integral]
+            #let iint = [#sym.integral.double]
+            #let infty = [#sym.infinity]
 
-#let di = [#math.dif i];
-#let du = [#math.dif u];
-#let dr = [#math.dif r];
-#let ds = [#math.dif s];
-#let dt = [#math.dif t];
-#let dx = [#math.dif x];
-#let dx = [#math.dif x];
-#let dy = [#math.dif y];
-#let dz = [#math.dif z];
+            #let mathbox(content) = {
+              style(styles => {
+                let size = measure(content, styles);
+                block(
+                    radius: 0.2em,
+                    stroke: fg,
+                    inset: size.height / 1.5,
+                    content)})
+            };
+        "#;
+        imports.to_string() + theme + page_size + text_options + math_remaining
+    }
+    fn add_source(&mut self, src: &str) {
+        self.source = Some(Source::detached(Self::preamble() + src))
+    }
+    fn revert(&mut self) {
+        self.source = None;
+    }
 
-#let int = [#sym.integral]
-#let infty = [#sym.infinity]
+    pub(crate) fn render(&mut self, msg: &str) -> Result<Vec<u8>, RenderErrors> {
+        self.add_source(msg);
+        let mut tracer = Tracer::default();
+        let document =
+            typst::compile(self, &mut tracer).map_err(|e| RenderErrors::SourceError(e.to_vec()))?;
+        let frame = document.pages.get(0).ok_or(RenderErrors::NoPageError)?;
+        let pixels = determine_pixels_per_point(frame.size())?;
 
-#let mathbox(content) = {
-  style(styles => {
-    let size = measure(content, styles);
-      block(
-        radius: 0.2em,
-        stroke: fg,
-        inset: size.height / 1.5,
-        content)})
-};"
+        let pixmap = typst_render::render(
+            frame,
+            pixels as f32,
+            typst::visualize::Color::from_u8(0, 0, 0, 0),
+        );
+
+        let mut writer = Cursor::new(Vec::new());
+
+        image::write_buffer_with_format(
+            &mut writer,
+            bytemuck::cast_slice(pixmap.pixels()),
+            pixmap.width(),
+            pixmap.height(),
+            image::ColorType::Rgba8,
+            image::ImageFormat::Png,
+        )
+        .unwrap();
+
+        let image = writer.into_inner();
+        self.revert();
+
+        Ok(image)
     }
 }
 
-const DESIRED_RESOLUTION: f32 = 2000.0;
-const MAX_SIZE: f32 = 10000.0;
-const MAX_PIXELS_PER_POINT: f32 = 15.0;
+fn determine_pixels_per_point(size: Axes<Abs>) -> Result<f64, RenderErrors> {
+    let desired_resolution = 2000.0;
+    let max_size = 10000.0;
+    let max_pixels_per_point = 15.0;
 
-pub(crate) fn determine_pixels_per_point(size: Size) -> Result<f32, RenderErrors> {
-    let x = size.x.to_pt() as f32;
-    let y = size.y.to_pt() as f32;
+    let x = size.x.to_pt();
+    let y = size.y.to_pt();
 
-    if x > MAX_SIZE || y > MAX_SIZE {
+    if x > max_size || y > max_size {
         Err(RenderErrors::PageSizeTooBig)
     } else {
         let area = x * y;
-        Ok((DESIRED_RESOLUTION / area.sqrt()).min(MAX_PIXELS_PER_POINT))
+        Ok((desired_resolution / area.sqrt()).min(max_pixels_per_point))
     }
 }
 
-pub(crate) struct ToCompile {
-    typst_essentials: Arc<TypstEssentials>,
-    source: Source,
-    time: time::OffsetDateTime,
-}
-
-fn string2source(source: String) -> Source {
-    Source::new(FileId::detached(), source)
-}
-
-impl ToCompile {
-    pub(crate) fn new(te: Arc<TypstEssentials>, source: String) -> Self {
-        ToCompile {
-            typst_essentials: te,
-            source: string2source(source),
-            time: time::OffsetDateTime::now_utc(),
-        }
-    }
-}
-
-impl typst::World for ToCompile {
-    fn book(&self) -> &Prehashed<FontBook> {
-        &self.typst_essentials.fontbook
-    }
-
-    fn file(&self, _id: FileId) -> FileResult<Bytes> {
-        Err(typst::diag::FileError::Other)
-    }
-
-    fn font(&self, id: usize) -> Option<Font> {
-        self.typst_essentials.fonts.get(id).cloned()
-    }
+impl typst::World for TypstRendered {
+    #[doc = " The standard library."]
+    #[doc = ""]
+    #[doc = " Can be created through `Library::build()`."]
     fn library(&self) -> &Prehashed<Library> {
-        &self.typst_essentials.library
+        &self.library
     }
 
+    #[doc = " Metadata about all known fonts."]
+    fn book(&self) -> &Prehashed<FontBook> {
+        &self.fontbook
+    }
+
+    #[doc = " Access the main source file."]
     fn main(&self) -> Source {
-        self.source.clone()
+        let Some(e) = &self.source else {
+            unreachable!()
+        };
+        e.clone()
     }
 
-    fn source(&self, _id: FileId) -> FileResult<Source> {
-        Ok(self.source.clone())
+    #[doc = " Try to access the specified source file."]
+    fn source(&self, id: FileId) -> FileResult<Source> {
+        if id == self.main().id() {
+            return Ok(self.main());
+        }
+        self.file_entry(&id)
+            .map_err(|_| FileError::NotSource)
+            .and_then(|f| f.clone().source(id))
     }
-    fn today(&self, offset: Option<i64>) -> Option<typst::eval::Datetime> {
-        let offset = offset.unwrap_or(0);
-        let offset = time::UtcOffset::from_hms(offset.try_into().ok()?, 0, 0).ok()?;
-        let time = self.time.checked_to_offset(offset)?;
-        Some(typst::eval::Datetime::Date(time.date()))
+
+    #[doc = " Try to access the specified file."]
+    fn file(&self, id: FileId) -> FileResult<Bytes> {
+        return self
+            .file_entry(&id)
+            .map_err(|_| FileError::NotSource)
+            .map(|f| f.clone().bytes);
+    }
+
+    #[doc = " Try to access the font with the given index in the font book."]
+    fn font(&self, index: usize) -> Option<Font> {
+        self.fonts.get(index).cloned()
+    }
+
+    #[doc = " Get the current date."]
+    #[doc = ""]
+    #[doc = " If no offset is specified, the local date should be chosen. Otherwise,"]
+    #[doc = " the UTC date should be chosen with the corresponding offset in hours."]
+    #[doc = ""]
+    #[doc = " If this function returns `None`, Typst\'s `datetime` function will"]
+    #[doc = " return an error."]
+    fn today(&self, offset: Option<i64>) -> Option<Datetime> {
+        None
     }
 }
-
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub(crate) enum RenderErrors {
-    SourceError(Vec<SourceError>),
+    SourceError(Vec<SourceDiagnostic>),
     NoPageError,
     PageSizeTooBig,
 }
@@ -225,21 +299,30 @@ pub(crate) enum RenderErrors {
 impl std::fmt::Display for RenderErrors {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
+            RenderErrors::SourceError(e) => {
+                let mut s = String::new();
+                let a = e.into_iter().fold(s, |acc, x| {
+                    let q = x
+                        .hints
+                        .clone()
+                        .into_iter()
+                        .fold(String::new(), |acc, e| acc + e.as_str());
+                    let ret = match x.severity {
+                        Severity::Error => acc + "Error:",
+                        Severity::Warning => acc + "Warning:",
+                    } + " "
+                        + x.message.as_str()
+                        + "\nHints (if any):"
+                        + q.as_str();
+                    return ret;
+                });
+                write!(f, "{a}")
+            }
             RenderErrors::NoPageError => {
                 write!(f, "No pages found...")
             }
             RenderErrors::PageSizeTooBig => {
                 write!(f, "Page too big...")
-            }
-            RenderErrors::SourceError(err) => {
-                write!(
-                    f,
-                    "{}",
-                    err.iter()
-                        .fold(String::from("Syntax error(s):\n"), |acc, se| acc
-                            + se.message.as_str()
-                            + "\n")
-                )
             }
         }
     }
@@ -247,5 +330,11 @@ impl std::fmt::Display for RenderErrors {
 
 impl std::error::Error for RenderErrors {}
 
-pub(crate) static TYPST_BASE: Lazy<Arc<TypstEssentials>> =
-    Lazy::new(|| Arc::new(TypstEssentials::new()));
+fn typst_init() -> &'static Mutex<TypstRendered> {
+    static TYPST: OnceLock<Mutex<TypstRendered>> = OnceLock::new();
+    TYPST.get_or_init(|| Mutex::new(TypstRendered::new()))
+}
+
+pub(crate) async fn typst_render(msg: &str) -> Result<Vec<u8>, RenderErrors> {
+    typst_init().lock().await.render(msg)
+}
